@@ -34,7 +34,7 @@
     "qr.js": "0.0.0"
   },
   "peerDependencies": {
-    "react": "^15.5.3 || ^16.0.0 || ^17.0.0"
+    "react": "^16.3.0 || ^17.0.0"
   },
   "devDependencies": {
     "@babel/cli": "^7.13.14",
